+++ src/redux/actions/costperuse/clear-cost-per-use-data.js
@@ -1,0 +1,5 @@
+export const CLEAR_COST_PER_USE_DATA = 'CLEAR_COST_PER_USE_DATA';
+
+export const clearCostPerUseData = () => ({
+  type: CLEAR_COST_PER_USE_DATA,
+});