+++ src/redux/actions/usage-consolidation/get-usage-consolidation-key.js
@@ -1,0 +1,6 @@
+export const GET_USAGE_CONSOLIDATION_KEY = 'GET_USAGE_CONSOLIDATION_KEY';
+
+export const getUsageConsolidationKey = credentialId => ({
+  type: GET_USAGE_CONSOLIDATION_KEY,
+  payload: credentialId,
+});