+++ src/redux/actions/usage-consolidation/get-usage-consolidation.js
@@ -1,0 +1,6 @@
+export const GET_USAGE_CONSOLIDATION = 'GET_USAGE_CONSOLIDATION';
+
+export const getUsageConsolidation = credentialId => ({
+  type: GET_USAGE_CONSOLIDATION,
+  payload: credentialId,
+});